Moving processors outside of exported package
