+++ lib/scala/interpreter-dsl/src/main/java/org/enso/interpreter/dsl/impl/MethodProcessor.java
@@ -1,4 +1,4 @@
-package org.enso.interpreter.dsl;
+package org.enso.interpreter.dsl.impl;
 
 import java.io.IOException;
 import java.io.PrintWriter;
